+++ sdk/spring/spring-cloud-azure-resourcemanager/src/main/java/com/azure/spring/cloud/resourcemanager/implementation/crud/ResourceCrud.java
@@ -8,9 +8,10 @@ package com.azure.spring.cloud.resourcemanager.implementation.crud;
  *
  * @param <T> Azure resource type
  * @param <K> Azure resource key type
+ * @param <P> Azure resource properties.
  *
  */
-public interface ResourceCrud<T, K> {
+public interface ResourceCrud<T, K, P> {
 
     /**
      * Gets a value for a given key.
@@ -37,10 +38,28 @@ public interface ResourceCrud<T, K> {
     T create(K key);
 
     /**
+     * Creates a value for a given key and properties.
+     *
+     * @param key The key.
+     * @param properties The properties for the Azure resource creation.
+     * @return The created value.
+     */
+    T create(K key, P properties);
+
+    /**
      * Gets or creates a value for a given key.
      *
      * @param key The key.
      * @return The retrieved or created value.
      */
     T getOrCreate(K key);
+
+    /**
+     * Gets or creates a value for a given key and properties.
+     *
+     * @param key The key.
+     * @param properties The properties for the Azure resource creation.
+     * @return The retrieved or created value.
+     */
+    T getOrCreate(K key, P properties);
 }